refactor web server
